Move encoding from auth module rest client
