+++ client/tx/encode.go
@@ -1,13 +1,13 @@
-package rest
+package tx
 
 import (
 	"encoding/base64"
+	"io/ioutil"
 	"net/http"
 
 	"github.com/cosmos/cosmos-sdk/client/context"
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/types/rest"
-
 	"github.com/cosmos/cosmos-sdk/x/auth"
 )
 
@@ -19,28 +19,35 @@ type encodeResp struct {
 	Tx string `json:"tx"`
 }
 
-// EncodeTxRequestHandlerFn returns the encode tx REST handler.  In particular, it takes a
-// json-formatted transaction, encodes it to the Amino wire protocol, and responds with
-// base64-encoded bytes
+// EncodeTxRequestHandlerFn returns the encode tx REST handler. In particular,
+// it takes a json-formatted transaction, encodes it to the Amino wire protocol,
+// and responds with base64-encoded bytes.
 func EncodeTxRequestHandlerFn(cdc *codec.Codec, cliCtx context.CLIContext) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		var m encodeReq
-		// Decode the transaction from JSON
-		if ok := unmarshalBodyOrReturnBadRequest(cliCtx, w, r, &m); !ok {
+		var req encodeReq
+
+		body, err := ioutil.ReadAll(r.Body)
+		if err != nil {
+			rest.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
 			return
 		}
 
-		// Re-encode it to the wire protocol
-		txBytes, err := cliCtx.Codec.MarshalBinaryLengthPrefixed(m.Tx)
+		err = cdc.UnmarshalJSON(body, &req)
+		if err != nil {
+			rest.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
+			return
+		}
+
+		// re-encode it via the Amino wire protocol
+		txBytes, err := cliCtx.Codec.MarshalBinaryLengthPrefixed(req.Tx)
 		if err != nil {
 			rest.WriteErrorResponse(w, http.StatusInternalServerError, err.Error())
 			return
 		}
 
-		// Encode the bytes to base64
+		// base64 encode the encoded tx bytes
 		txBytesBase64 := base64.StdEncoding.EncodeToString(txBytes)
 
-		// Write it back
 		response := encodeResp{Tx: txBytesBase64}
 		rest.PostProcessResponse(w, cdc, response, cliCtx.Indent)
 	}
