+++ x/auth/client/rest/query.go
@@ -19,18 +19,11 @@ func RegisterRoutes(cliCtx context.CLIContext, r *mux.Router, cdc *codec.Codec, 
 		"/auth/accounts/{address}",
 		QueryAccountRequestHandlerFn(storeName, cdc, context.GetAccountDecoder(cdc), cliCtx),
 	).Methods("GET")
+
 	r.HandleFunc(
 		"/bank/balances/{address}",
 		QueryBalancesRequestHandlerFn(storeName, cdc, context.GetAccountDecoder(cdc), cliCtx),
 	).Methods("GET")
-	r.HandleFunc(
-		"/tx/broadcast",
-		BroadcastTxRequestHandlerFn(cdc, cliCtx),
-	).Methods("POST")
-	r.HandleFunc(
-		"/tx/encode",
-		EncodeTxRequestHandlerFn(cdc, cliCtx),
-	).Methods("POST")
 }
 
 // query accountREST Handler
